requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pyjwt==2.8.0
pydantic==2.5.3
aiosqlite==0.19.0
httpx==0.26.0
